+++ NEWS.md
@@ -7,6 +7,7 @@
 * `frequencies()`, `crosstabs()`: Now also capture the meta styling information. (01.09.2026)
 * `any_table()`, `frequencies()`, `crosstabs()`, `export_with_style()`: New `output` option "no" prevents any styling and additionally sets `print = FALSE` so that only the table and meta information are captured. (01.09.2026)
 * `combine_into_workbook()`: Can now also combine `frequencies()` and `crosstabs()` outputs. (01.09.2026)
+* `run_script()`, `run_folder()`, `run_project()`, `run_project_parallel()`: Single files and entire folders can be silenced (code will not be executed) by adding a "_" to the end of the file or folder name. (03.09.2026)
 
 ### Changed functionality
 
@@ -29,6 +30,8 @@
 * `recode_multi()`: When recoding a variable into numeric values then the desired numeric values are output instead of the factor values. (01.09.2026)
 * `recode_multi()`: When a variable is invalid then the function aborts with an error message instead of just crashing. (01.09.2026)
 * `if.()`, `else_if.()`, `else.()`, `ifelse_multi()`, `where.()`: When a variable inside a condition or an assignment is invalid then the function aborts with an error message instead of just crashing. (01.09.2026)
+* `import_data()`, `import_multi()`: Should now detect csv and txt encoding correctly. (03.09.2026)
+* `export_with_style()`: First column isn't set up as row header column anymore it is now treated as part of the table. (03.09.2026)
 
 ### Optimization